+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Auto switch to admin mode if admin on production
 - Updated supabase places types according to phacto specification 
 - Updated admin places tab names to "Places" and "Unverified"
+- Updated AGAM to use average rating if review_count > 0 otherwise fallback to gmaps_rating
 
 ### Fixed
 - Fixed upload button text stays as replace photo even after uploading a new photo